+++ segment_reader.go
@@ -19,8 +19,10 @@ import (
 //	r, err := NewSegmentReader(…)
 //	…
 //
-//	for r.Next() {
-//	  e, offset, err := r.Read()
+//	for r.ReadNext() {
+//	  offset := r.Offset()
+//	  …
+//	  entry, err := r.Decode()
 //	  …
 //	}
 //
@@ -48,14 +50,24 @@ func NewSegmentReader(r io.Reader, registry *EntryRegistry) (*SegmentReader, err
 	}, nil
 }
 
-// Next loads the data for the next Entry from the underlying reader.
+// SeekEnd reads through the entire segment until the end and returns the last offset.
+func (r *SegmentReader) SeekEnd() (lastOffset uint32, err error) {
+	for r.ReadNext() {
+		lastOffset = r.Offset()
+	}
+
+	return lastOffset, r.Err()
+}
+
+// ReadNext loads the data for the next Entry from the underlying reader.
 // For efficiency reasons, this function neither checks the entry checksum,
 // nor does it decode the entry bytes. This is done, so the caller can quickly
 // seek through a WAL up to a specific offset without having to decode each WAL
 // entry.
 //
-// In order to actually decode the read WAL entry, you need to use SegmentReader.Read(…).
-func (r *SegmentReader) Next() bool {
+// You can get the offset of the current entry using SegmentReader.Offset().
+// In order to actually decode the read WAL entry, you need to use SegmentReader.Decode(…).
+func (r *SegmentReader) ReadNext() bool {
 	var header [9]byte // 4B offset + 1B type + 4B checksum
 	n, err := io.ReadFull(r.r, header[:])
 	if err == io.EOF {
@@ -86,27 +98,31 @@ func (r *SegmentReader) Next() bool {
 	return true
 }
 
-// Read decodes the next Entry and returns it together with its WAL offset.
-// Before any entry can be read, the caller must call SegmentReader.Next() first.
-func (r *SegmentReader) Read() (entry Entry, offset uint32, err error) {
+// Offset returns the offset of the last entry that was read by SegmentReader.ReadNext().
+func (r *SegmentReader) Offset() uint32 {
+	return r.offset
+}
+
+// Decode decodes the last entry that was read using SegmentReader.ReadNext().
+func (r *SegmentReader) Decode() (Entry, error) {
 	if r.err != nil {
-		return nil, 0, r.err
+		return nil, r.err
 	}
 
 	if r.entry == nil {
-		return nil, r.offset, errors.New("must call SegmentReader.Next() first")
+		return nil, errors.New("must call SegmentReader.ReadNext() first")
 	}
 
 	if r.checksum != crc32.ChecksumIEEE(r.payload) {
-		return nil, r.offset, fmt.Errorf("detected WAL Entry corruption at WAL offset %d", r.offset)
+		return nil, fmt.Errorf("detected WAL Entry corruption at WAL offset %d", r.offset)
 	}
 
-	err = r.entry.DecodePayload(r.payload)
-	return r.entry, r.offset, err
+	err := r.entry.DecodePayload(r.payload)
+	return r.entry, err
 }
 
-// Err returns any error that happened when calling Next(). This function must
-// always be called even if Next() never returned true.
+// Err returns any error that happened when calling ReadNext(). This function must
+// always be called even if ReadNext() never returned true.
 //
 // Please refer to the comment on the SegmentReader type to see the full usage pattern.
 func (r *SegmentReader) Err() error {