+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Introduce `EntryRegistry` to improve code readability
+- Introduce `EntryRegistry` to improve code readability (fgrosse/wal#3)
+- Introduce `github.com/fgrosse/wal/waltest` package (fgrosse/wal#5)
+- Refactor `SegmentReader` API for better performance and readability (fgrosse/wal#5)
 
 ## [v0.1.0] - 2023-03-12
 - Initial release